Added empty analysis fragment - Kyle Barry (u19232510)
