+++ mobile/app/build.gradle
@@ -60,6 +60,7 @@ dependencies {
     implementation 'androidx.navigation:navigation-ui-ktx:2.3.0'
     implementation 'androidx.annotation:annotation:1.2.0'
     implementation 'com.google.devtools.ksp:symbol-processing-api:1.5.20-1.0.0-beta04'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     //kapt 'groupId:artifactId:version'
     testImplementation 'junit:junit:4.+'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
