--- mobile/app/src/main/java/com/bdpsolutions/highfive/data/login/model/LoggedInUser.kt
@@ -1,9 +1,0 @@
-package com.bdpsolutions.highfive.data.login.model
-
-/**
- * Data class that captures user information for logged in users retrieved from LoginRepository
- */
-data class LoggedInUser(
-    val userId: String,
-    val displayName: String
-)
