+++ mobile/app/src/main/res/values/strings.xml
@@ -13,4 +13,5 @@
     <string name="invalid_username">Not a valid username</string>
     <string name="invalid_password">Password must be >5 characters</string>
     <string name="login_failed">"Login failed"</string>
+    <string name="analysis_fragment">Analysis</string>
 </resources>
